refactor: colocate column presentation helpers
