+++ src/columns/determinate.ts
@@ -1,4 +1,4 @@
-import type { TaskSnapshot } from "../../task-model";
+import type { TaskSnapshot } from "../task-model";
 
 export const isDeterminate = (
   task: TaskSnapshot,
