+++ src/columns/elapsed-eta.tsx
@@ -1,7 +1,7 @@
 import type { ColumnDef } from "./types";
 import type { TaskSnapshot } from "../task-model";
 import { Text } from "ink";
-import { formatElapsedEta } from "../renderer/shared/format";
+import { formatElapsedEta } from "./format";
 import { useNow } from "../renderer/context/now-context";
 
 const ElapsedEtaCell = ({ task }: { readonly task: TaskSnapshot }) => {
