+++ src/columns/elapsed.tsx
@@ -1,7 +1,7 @@
 import type { ColumnDef } from "./types";
 import type { TaskSnapshot } from "../task-model";
 import { Text } from "ink";
-import { formatElapsed } from "../renderer/shared/format";
+import { formatElapsed } from "./format";
 import { useNow } from "../renderer/context/now-context";
 
 const ElapsedCell = ({ task }: { readonly task: TaskSnapshot }) => {
