+++ src/terminal/text-width.ts
