+++ tests/columns/format.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import * as Progress from "../../src";
 import { getTaskIndicator } from "../../src/columns/description";
-import { formatAmount } from "../../src/renderer/shared/format";
+import { formatAmount } from "../../src/columns/format";
 
 const makeTask = (
   units: Progress.TaskSnapshot["units"],
